requirements.txt
python-can
canopen
urwid